[Feat] Move Chains enum into src/types.
