+++ src/machines/signing/definition.ts
@@ -1,5 +1,5 @@
 import { assign, createMachine } from "xstate";
-import { Chains } from "src/utils/constants";
+import { Chains } from "src/types";
 
 const defaultContext = {
   dapp: {
